[FAB2023-999] refactor: 데이터 모델 추가 모달 개발 - query 관련 기능 composable로 이동 및 적용
